+++ lib/tiny_rest_client/response.rb
@@ -3,13 +3,14 @@
 # The Response class represents a response object that
 # can be used to represent success or failure outcomes
 class Response
-  attr_reader :code, :status, :body, :headers
+  attr_reader :code, :status, :body, :headers, :timed_out
 
-  def initialize(code: nil, status: nil, body: nil, headers: nil)
+  def initialize(code: nil, status: nil, body: nil, headers: nil, timed_out: false)
     @code = code&.to_i
     @status = status
     @headers = headers
     @body = parse_body(body)
+    @timed_out = timed_out
   end
 
   def success?
@@ -19,6 +20,12 @@ class Response
   def failure?
     !success?
   end
+
+  def timed_out?
+    timed_out
+  end
+
+  private
 
   def parse_body(raw_body)
     return raw_body unless raw_body.is_a?(String)